Pass varied parameter via learner option
The ALPrequentialEvaluationTask does not need an own option that
is simply passed to its learner on instantiation. Therefore, the
learner option passed to the ALPrequentialEvaluationTask already
contains all specified parameters for the learner. This also works
better with arbitrarily chosen parameters, not only budget.

In refreshBudgetParamNameOption, the last learner option is cached
in order to check, if an update is actually needed. Otherwise, the
value of the EditableMultiChoiceOption could be resetted too oft