Fix scaling on window resize
When reducing the window size, the canvas class did not properly resize. This was caused by updating the size of the canvas children (axesPanel and its children curvePanel/scatterPanel and eventPanel) by overriding the paintComponent() method. paintComponent() however is only called if a component is not obscured by its children (see also JComponent.paint()), which is the case if the window size is reduced (the canvas size is recomputed automatically, but the children's size not).

This is fixed by overriding paintChildren() instead of paintComponent(), as paintChildren() is not bound to the !rectangleIsObscured() condition.