Wrap options of subtasks instead of copying
The ALCrossValidationTask and the ALMultiParamTask used to copy all the options
of their subtasks and passed the configured values. As this makes the
configuration windows too big and does not adapt to changes of the subtask
layout, it was changed so that the subtasks can be configured in separate
windows that can be reached via a ClassOption.