Implement proper measure overview
The class moa.gui.active.MeasureOverview replaces the ClusteringVisualEvalPanel to display a live preview of the measures. Instead of displaying the current and mean values of the current and last task, it only displays the current value of the current task. Additionally, it provides a JComboBox to switch between different varying parameters, if they are available.