Rename ThresholdBM to FixedBM and clarify budget
Previously, a threshold of 0.1 meant that for each random number
above 0.1 an instance may be aquired. Therefore, the actually used
budget was 0.9. As this was not consistent with the use of budget
parameters in other learners and budget managers, the TresholdBM has
been changed to the FixedBM that specifies a fixed budget and
enforces it exactly.