Create class MeanPreviewCollection
This class holds separate PreviewCollections for mean values and
standard deviations. This way, standard deviation values do not
need to be appended to the mean measurement arrays, but can stay
well separated.
Be aware that the current state of the branch is not working,
because the ALTaskTextViewerPanel needs to be adapted to these
changes.